refactor(web): name browser modules' files after the exports they carry
Every other browser/web-path file in the repo already carries its
runtime prefix; after the symbol rename, src/web's bare file names were
the odd ones out and no longer matched their primary exports. main.tsx
keeps its bundler-conventional name.
