+++ AGENTS.md
@@ -38,8 +38,8 @@ ReviewIntent + caller facts -> planReviewIntent -> ReviewAction[] -> reducer -> 
   operations start as intents. Callers supply mutable-note IDs/timestamps; core derives identities.
 - **Surfaces/publishers:** `useTerminalReview.ts` is the TUI adapter and
   `reviewNoteMapping.ts` is terminal-only; `src/web/` is the browser client — read-only today
-  — where `reviewApiClient`/`reviewMirror` speak the Phase 4 HTTP contracts and
-  `pierreDocument`/`BrowserReviewStream` render with Pierre. Rows, measurement, scrolling, layout,
+  — where `browserReviewApiClient`/`browserReviewMirror` speak the Phase 4 HTTP contracts and
+  `browserPierreDocument`/`BrowserReviewStream` render with Pierre. Rows, measurement, scrolling, layout,
   themes, DOM mechanics, and source I/O stay local to each surface. `useHunkSessionBridge.ts` publishes the current terminal
   session export; `registration.ts` builds its metadata/initial snapshot and `bridge.ts` receives
   agent commands. This broker export is not a full `ReviewState` mirror.
