+++ src/web/BrowserReviewStream.test.tsx
@@ -5,11 +5,11 @@ import { formatReviewAddress } from "../core/review/address";
 import { projectReviewDocument } from "../core/review/document";
 import { createTestDiffFile, createTestSourceFetcher } from "../../test/helpers/diff-helpers";
 import { reviewErrorMessage } from "../session/reviewErrorCatalog";
-import { buildBrowserReviewFileRenderModel } from "./pierreDocument";
+import { buildBrowserReviewFileRenderModel } from "./browserPierreDocument";
 import { reviewHttpFailure } from "../session/reviewHttpProtocol";
-import type { BrowserReviewSourceEntry } from "./reviewSources";
-import { BrowserGapStrip, BrowserReviewStream } from "./ReviewStream";
-import { DEFAULT_BROWSER_VIEW_OPTIONS } from "./viewOptions";
+import type { BrowserReviewSourceEntry } from "./browserReviewSources";
+import { BrowserGapStrip, BrowserReviewStream } from "./BrowserReviewStream";
+import { DEFAULT_BROWSER_VIEW_OPTIONS } from "./browserViewOptions";
 
 const BASE = `${Array.from({ length: 24 }, (_unused, index) => `line ${index + 1}`).join("\n")}\n`;
 const CHANGED = BASE.replace("line 4", "line 4 changed");
