+++ src/web/browserReviewMirror.ts
@@ -47,7 +47,7 @@ import type { ReviewDocumentV1, ReviewFileV1 } from "../core/review/types";
 import { reviewErrorMessage } from "../session/reviewErrorCatalog";
 import { reviewHttpFailure, type HunkReviewPublicationBodyV1 } from "../session/reviewHttpProtocol";
 import type { HunkReviewResourceCatalogV1 } from "../session/reviewProtocol";
-import type { BrowserReviewApiClient, BrowserReviewFailure } from "./reviewApiClient";
+import type { BrowserReviewApiClient, BrowserReviewFailure } from "./browserReviewApiClient";
 
 /**
  * How soon a dropped stream is retried, and how far apart retries grow.
